% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-patterns.R
\name{create_pattern_id_from_rgba_vec}
\alias{create_pattern_id_from_rgba_vec}
\title{Create a pattern ID}
\usage{
create_pattern_id_from_rgba_vec(pattern_pkg, rgba_vec, ...)
}
\arguments{
\item{pattern_pkg}{name of the package which contains the pattern handling}

\item{rgba_vec}{colour vector. elements in [0, 255]}

\item{...}{other arguments to foreign \code{create_pattern_id()} call}
}
\description{
Create a pattern ID
}
