% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-style.R
\name{fill_style}
\alias{fill_style}
\title{Construct a vector of styles related to 'fill'}
\usage{
fill_style(state, use_pattern_fill)
}
\arguments{
\item{state}{list including 'gc' (graphics context)}

\item{use_pattern_fill}{whether or not to use the pattern fill for this style}
}
\value{
character vector of multiple "{attr_name}: {value}" strings for 'fill',
        'fill-opacity'
}
\description{
Construct a vector of styles related to 'fill'
}
