% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-style.R
\name{filter_style}
\alias{filter_style}
\title{Construct a vector of styles related to 'filter'}
\usage{
filter_style(state, geom)
}
\arguments{
\item{state}{list including 'gc' (graphics context)}

\item{geom}{which geometry has asked for a style}
}
\value{
character  "{attr_name}: {value}" strings for 'filter' or NULL
        if no filter available.
}
\description{
Construct a vector of styles related to 'filter'
}
