% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glue_two}
\alias{glue_two}
\title{Custom version of \code{glue} which rounds all numerics to 2 decimals}
\usage{
glue_two(..., .envir = parent.frame())
}
\arguments{
\item{..., .envir}{standard glue args}
}
\value{
interpolated string
}
\description{
Custom version of \code{glue} which rounds all numerics to 2 decimals
}
