% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{is_black}
\alias{is_black}
\title{Is the colour black?}
\usage{
is_black(vec)
}
\arguments{
\item{vec}{4 element rgba with integer values [0, 255]}
}
\value{
TRUE if black
}
\description{
Is the colour black?
}
