% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{is_transparent}
\alias{is_transparent}
\title{Is the colour transparent?}
\usage{
is_transparent(vec)
}
\arguments{
\item{vec}{4 element rgba with integer values [0, 255]}
}
\value{
TRUE if alpha == 0
}
\description{
Is the colour transparent?
}
