% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{is_white}
\alias{is_white}
\title{Is the colour white?}
\usage{
is_white(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element rgba with integer values [0, 255]}
}
\value{
TRUE if white
}
\description{
Is the colour white?
}
