% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{rgba_to_alpha}
\alias{rgba_to_alpha}
\title{Calcuate the alpha value from the 4-element RGBA vector}
\usage{
rgba_to_alpha(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element rgba with integer values [0, 255]}
}
\value{
alpha in range [0,1]
}
\description{
Calcuate the alpha value from the 4-element RGBA vector
}
