% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{rgba_to_hex}
\alias{rgba_to_hex}
\title{Convert a 4-element rgba vector to hex RGB}
\usage{
rgba_to_hex(rgba_vec)
}
\arguments{
\item{rgba_vec}{4 element rgba with integer values [0, 255]}
}
\value{
colour hex triplet as string e.g. "#0156fe"
}
\description{
Convert a 4-element rgba vector to hex RGB
}
