% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-style.R
\name{stroke_style}
\alias{stroke_style}
\title{Construct a vector of styles related to 'stroke'}
\usage{
stroke_style(state)
}
\arguments{
\item{state}{list including 'gc' (graphics context)}
}
\value{
character vector of multiple "{attr_name}: {value}" strings for
        'stroke', 'stroke-width', 'stroke-opacity'
}
\description{
Note: 1 lwd = 1/96", but units in rest of document are 1/72"
}
