% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgout.R
\name{svg_callback}
\alias{svg_callback}
\title{The main SVG callback.}
\usage{
svg_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device call}

\item{args}{arguments to the call}

\item{state}{rdata, gc and dd}
}
\description{
The main SVG callback.
}
