% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgout.R
\name{svg_strWidth}
\alias{svg_strWidth}
\title{Return the width of the given string}
\usage{
svg_strWidth(args, state)
}
\arguments{
\item{args, state}{standard pass-through from device driver}
}
\description{
Return the width of the given string
}
