% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgout.R
\name{svgout}
\alias{svgout}
\title{SVG device written in R. JUst sets up \code{rdevice} to call \code{svg_callback()}}
\usage{
svgout(filename = "svgout.svg", width = 6, height = 4, ...)
}
\arguments{
\item{filename}{default: "svgout.svg"}

\item{width, height}{size in inches}

\item{...}{arguments passed to \code{devout::rdevice}}
}
\description{
SVG device written in R. JUst sets up \code{rdevice} to call \code{svg_callback()}
}
