% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformer_round}
\alias{transformer_round}
\title{Transformer for \code{glue} to round numeric values to decimal places and convert NULL to text}
\usage{
transformer_round(text, envir)
}
\arguments{
\item{text, envir}{standard transformer arguments}
}
\value{
evaluated text value
}
\description{
Transformer for \code{glue} to round numeric values to decimal places and convert NULL to text
}
