% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{vec2alpha}
\alias{vec2alpha}
\title{Calcuate the alpha value from the 4-element RGBA vector}
\usage{
vec2alpha(vec)
}
\arguments{
\item{vec}{4 element rgba with integer values [0, 255]}
}
\value{
alpha in range [0,1]
}
\description{
Calcuate the alpha value from the 4-element RGBA vector
}
