% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{vec2rgb}
\alias{vec2rgb}
\title{Convert a 4-element rgba vector to hex RGB}
\usage{
vec2rgb(vec)
}
\arguments{
\item{vec}{4 element rgba with integer values [0, 255]}
}
\value{
colour hex triplet as string e.g. "#0156fe"
}
\description{
Convert a 4-element rgba vector to hex RGB
}
