% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-patterns.R
\name{write_pattern}
\alias{write_pattern}
\title{Write a pattern to the state.}
\usage{
write_pattern(state, ...)
}
\arguments{
\item{state}{device state}

\item{...}{other arguments to foreign \code{decode_pattern_from_rgba_vec()} call}
}
\value{
state
}
\description{
Write a pattern to the state.
}
