% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-style.R
\name{create_style_string}
\alias{create_style_string}
\title{Create stryle strings for the given attritute or set of attributes}
\usage{
create_style_string(attr_name, state, geom)
}
\arguments{
\item{attr_name}{attribute name of name for a set of attributes}

\item{state}{list including 'gc' (graphics context)}

\item{geom}{which geometry has asked for a style}
}
\value{
character vector of one or more "{attr_name}: {value}" strings
}
\description{
Create stryle strings for the given attritute or set of attributes
}
